improve the Keycloak test
Moved the application auth configuration from web.xml
to project-defaults.yml. This means the entirety of security
configuration is on one place.

Increased the timeout for starting the Keycloak container;
the default timeout of 10 seconds is too short, because starting
Keycloak easily takes just about that.

Moved the configuration of the Docker Maven plugin to the root
POM, so that the common configuration is only on one place.
Also using the Docker Maven plugin requires less code; this
will be nice when it's used in other modules.