improve Topology test stability
In addition to waiting for MSC stability, it is also required
to wait for topology listeners to be invoked. It would be possible
to just `Thread.sleep`, but instead, this commit uses Awaitility,
a nice DSL for awaiting events.

Technically, awaiting MSC stability could be removed after
introducing Awaitility, but let's keep that in hope it will help
understanding what's going on in the test.