improve topology tests so that running manually is possible
The `topology` and `topology-webapp` tests need custom dependencies
added to the deployment. Using `jboss-deployment-structure.xml`
isn't required in the tests, because there are other means of
specifying the dependencies, so `jboss-deployment-structure.xml`
wasn't included.

This commit adds the `jboss-deployment-structure.xml` files
to both `topology` and `topology-webapp` so that these modules
can be used manually, outside of a test.