fix AssertJ and Awaitility usage in topology test
The `TopologyTest` is an in-container Arquillian test, so it can't
really use AssertJ and Awaitility, because Arquillian doesn't bundle
them into the tested archive. For some reason, it used to work (maybe
WildFly Swarm added the JARs to the deployment?), but it doesn't now.

Fixed by adding a custom Arquillian extension. This extension creates
a `AuxiliaryArchiveAppender` which adds AssertJ and Awaitility.