use AssertJ in CDI + Bean Validation test
The `CdiBeanValidationTest` is an in-container Arquillian test, so
it can't really use AssertJ, because Arquillian doesn't bundle it
into the tested archive. The test used Hamcrest matchers with their
JUnit support, as Arquillian adds them by default, but we'd really
like to use AssertJ everywhere.

This commit moves the test to AssertJ, which is enabled by adding
a custom Arquillian extension. This extension creates
a `AuxiliaryArchiveAppender` which adds AssertJ.