always use uber-jars when starting tested applications
By default, the WildFly Swarm Maven plugin starts applications from
classpath. This, however, is very wrong in a lot of situations, and
can cause weird and hard to track issues. This commit explicitly
changes the configuration to always build an uber-jar and use it
for tests. This actually improves test coverage, because in real
world, applications are always started from uber-jars, and running
them from classpath was just an optimization for tests.