fix OpenTracing dependencies
This commit removes the `io.jaegertracing` specification in root
`<dependencyManagement>`, because it would win over transitive
dependencies of the `jaeger` fraction. That's wrong, because
in product, the fraction brings transitively the productized bits
of Jaeger, which is what we need.

Also this commit reorders the dependencies in `opentracing/basic`,
because there are 2 ways how `io.opentracing` dependencies could be
brought in transitively: through `io.jaegertracing` or through
the `opentracing` fraction. In product, we need the productized bits,
which are brought in through the `opentracing` fraction, so
the dependency on the fraction needs to go first.