[ENTSWM-714] enforce TLSv1.2 support in HTTP client in Elytron tests
This is to make these tests pass on the IBM JDK. The IBM JSSE provider
by default only enables TLSv1, and in the Elytron tests, we configure
the server SSL context to only enable TLSv1.2. Therefore, we have to
enforce TLSv1.2 on the client side.

Alternatively, we could configure the IBM JSSE provider to match
the SunJSSE provider behavior (used in OpenJDK) by setting
the `com.ibm.jsse2.overrideDefaultTLS` system property to `true`.
In this mode, the IBM JSSE provider by default enables TLSv1,
TLSv1.1 and TLSv1.2. It would be easier (simple change in Maven
Surefire and Failsafe plugins configuration), but if we configure
the server to only enable TLS v1.2, it's probably more instructive
to configure the client in the same way.

The relevant IBM JDK documentation is:
- https://www.ibm.com/support/pages/how-do-i-change-default-ssl-protocol-my-java-client-application-will-use
- https://www.ibm.com/support/knowledgecenter/en/SSYKE2_7.1.0/com.ibm.java.security.component.71.doc/security-component/jsse2Docs/matchsslcontext_tls.html