fix servlet line separator expectations in tests (#99)
A lot of tests use servlet on the application side and write
a response using `println`. That previously always used `\n`
as a line separator, but UNDERTOW-1792 changed that to `\r\n`.

Considering Thorntail's lifecycle, I think it's fine to just
change all the tests to expect `\r\n` instead of `\n`.